First changes to clean up the project after the fork
Added Tailwind CSS as the new CSS library, no translation of the
current CSS has been done yet.

Added the components(Bracket, Round, Matchup, Player) from the brackets
branch on the original repository.

Added craco to the project to support tailwind. The new scripts are
craco start, craco build and so on.