Updated the Home page styles to use Tailwind
I made the buttons reactive by storing the state of the mouse hover
on the buttons.

Created a default-button class with tailwind for consistency of buttons
across the site.

There are still a few default CSS classes on the home page that need to
be changed, this is what the next commit will be.