Updated styling for the header (Header.js)
Created a header link CSS class to be used for all header links.

Switched the Team Seeder h1 to a link to the home page and
removed the link to "Home" in the header.

Made the root div in App.js a flex box to properly size
the elements on the screen so that they fill the view.